+++ src/NServiceBus.Hosting.Tests/APIApprovals.cs
@@ -1,4 +1,4 @@
-﻿using NServiceBus.Hosting.Windows;
+﻿using NServiceBus;
 using NUnit.Framework;
 using Particular.Approvals;
 using PublicApiGenerator;
@@ -9,7 +9,7 @@ public class APIApprovals
     [Test]
     public void Approve()
     {
-        var publicApi = ApiGenerator.GeneratePublicApi(typeof(Program).Assembly);
+        var publicApi = ApiGenerator.GeneratePublicApi(typeof(HostingInternalType).Assembly);
         Approver.Verify(publicApi);
     }
 }
