+++ src/NServiceBus.Hosting.Windows/AssemblyInfo.cs
@@ -2,4 +2,4 @@
 using System.Runtime.InteropServices;
 
 [assembly: ComVisible(false)]
-//[assembly: CLSCompliant(true)]
+[assembly: CLSCompliant(false)]
