--- src/NServiceBus.Hosting.Windows/Content/Help.txt
@@ -1,36 +1,0 @@
-﻿NServiceBus Endpoint Host Service
-
-USAGE:
-   NServiceBus.Host.exe [-install] [options]
-   NServiceBus.Host.exe [-uninstall] [options]
-
-INSTALL OPTIONS:
-
-{0}
-
-UNINSTALL OPTIONS:
-
-{1}
-
-If no service name is specified NServiceBus will use the endpoint name along with the version number of the assembly it is contained within, for example:
-
-   MyProject.MyEndpoint-1.0.0
-
-The default for the display name is the same value as the service name, and the description defaults to a generic NServiceBus host description.
-
-You can also specify the endpoint configuration type in the file NServiceBus.Host.exe.config. This file is optional.
-
-If you don't specify the endpoint configuration type either in the command-line or in the NServiceBus.Host.exe.config file, all the DLLs in the runtime directory will be scanned for a type that implements NServiceBus.IConfigureThisEndpoint.
-
-If you set the service name and sidebyside during installation you will need to specify them when uninstalling them as well, eg:
-
-   NServiceBus.Host.exe -uninstall -serviceName="MyPublisher" -sidebyside
-
-EXAMPLES:
-   NServiceBus.Host.exe -install
-     -serviceName="MyPublisher"
-     -displayName="My Publisher Service"
-     -description="Service for publishing event messages"
-     -endpointConfigurationType="YourEndpointConfigType.YourNameSpace, YourAssembly"
-     -username="corp\serviceuser"
-     -password="p@ssw0rd!" NServiceBus.Production
