--- src/NServiceBus.Hosting.Windows/EndpointConfig.cs.pp
@@ -1,23 +1,0 @@
-
-namespace $rootnamespace$
-{
-    using NServiceBus;
-
-    public class EndpointConfig : IConfigureThisEndpoint
-    {
-        public void Customize(EndpointConfiguration endpointConfiguration)
-        {
-            //TODO: NServiceBus provides multiple durable storage options, including SQL Server, RavenDB, and Azure Storage Persistence.
-            // Refer to the documentation for more details on specific options.
-            endpointConfiguration.UsePersistence<PLEASE_SELECT_ONE>();
-
-            // NServiceBus will move messages that fail repeatedly to a separate "error" queue. We recommend
-            // that you start with a shared error queue for all your endpoints for easy integration with ServiceControl.
-            endpointConfiguration.SendFailedMessagesTo("error");
-
-            // NServiceBus will store a copy of each successfully process message in a separate "audit" queue. We recommend
-            // that you start with a shared audit queue for all your endpoints for easy integration with ServiceControl.
-            endpointConfiguration.AuditProcessedMessagesTo("audit");
-        }
-    }
-}
