+++ src/NServiceBus.Hosting.Windows/InternalsVisibleTo.cs
@@ -1,4 +1,3 @@
 ﻿using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("NServiceBus.Hosting.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100dde965e6172e019ac82c2639ffe494dd2e7dd16347c34762a05732b492e110f2e4e2e1b5ef2d85c848ccfb671ee20a47c8d1376276708dc30a90ff1121b647ba3b7259a6bc383b2034938ef0e275b58b920375ac605076178123693c6c4f1331661a62eba28c249386855637780e3ff5f23a6d854700eaa6803ef48907513b92")]
-[assembly: InternalsVisibleTo("NServiceBus.Hosting.ComponentTests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100cd8d2c941187b37b1108e7b07d88a7af9b6b3e133152f3d5403f0dba54a2a0c22111ace29f883c5a8e3899a3e150586cc332aa86a4a730c45acf07afacb38a400f5b55de677c60c54fae957d4af7f6986af56c668d33bf0c49c5360dba868a4ae8e297d1dec659a5eb6e97da27aa944edda1dc33d83824e3b0ee9df4dc1747dc")]
