--- src/NServiceBus.Hosting.Windows/Profiles/Handlers/IntegrationProfileHandler.cs
@@ -1,13 +1,0 @@
-﻿namespace NServiceBus.Hosting.Windows.Profiles.Handlers
-{
-
-    using Hosting.Profiles;
-
-    class IntegrationProfileHandler : IHandleProfile<Integration>
-    {
-        public void ProfileActivated(EndpointConfiguration config)
-        {
-            config.EnableInstallers();
-        }
-    }
-}
