--- src/NServiceBus.Hosting.Windows/Profiles/IProfile.cs
@@ -1,10 +1,0 @@
-﻿namespace NServiceBus
-{
-    /// <summary>
-    /// Marker interface to indicate a run-time profile.
-    /// Implementors must be concrete class - interfaces are not supported.
-    /// </summary>
-    public interface IProfile
-    {
-    }
-}
