--- src/NServiceBus.Hosting.Windows/Profiles/Integration.cs
@@ -1,9 +1,0 @@
-﻿namespace NServiceBus
-{
-    /// <summary>
-    /// Indicates that infrastructure suitable for integration environments be used.
-    /// </summary>
-    public interface Integration : IProfile
-    {
-    }
-}
