--- src/NServiceBus.Hosting.Windows/Profiles/Lite.cs
@@ -1,9 +1,0 @@
-﻿namespace NServiceBus
-{
-    /// <summary>
-    /// Indicates that the lightest weight infrastructure should be used.
-    /// </summary>
-    public interface Lite : IProfile
-    {
-    }
-}
