--- src/NServiceBus.Hosting.Windows/Profiles/Production.cs
@@ -1,9 +1,0 @@
-﻿namespace NServiceBus
-{
-    /// <summary>
-    /// Indicates that the infrastructure should configure itself for production.
-    /// </summary>
-    public interface Production : IProfile
-    {
-    }
-}
